requirements.txt
fastapi
fastapi-cors
uvicorn
openai
python-dotenv
gtts
audiorecorder
requests
python-multipart
pytest
streamlit